third_party/blink/web_tests/fast/html/imports/rel-import-to-style-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
Updating attribute value shouldn't invalidate loaded import.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Before update:
PASS target.sheet is null
PASS target.import is non-null.
After update:
PASS target.sheet is null
PASS target.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
